tests: Units tests created by service

+++ src/services/DeleteEmployeeService.ts
@@ -1,5 +1,6 @@
 import { injectable, inject } from 'tsyringe';
 import Employee from '../entities/Employee';
+import AppError from '../errors/AppError';
 import IEmployeeRepository from '../repositories/IEmployeeRepository';
 
 @injectable()
@@ -10,7 +11,11 @@ export default class DeleteEmployeeService {
   ) {}
 
   public async execute(CPF: string): Promise<Employee | Employee[]> {
+    if (!CPF) throw new AppError('CPF must be required');
     const result = await this.employeeRepository.remove(CPF);
+
+    if (Array.isArray(result) && result.length === 0)
+      throw new AppError('Employees not found.', 404);
     return result;
   }
 }
